+++ src/main/java/com/log_statistics_service/domain/NextLogResult.java
@@ -1,5 +1,5 @@
 package com.log_statistics_service.domain;
 
 
-public record NextLogResult(String nextLog, int offset) {
+public record NextLogResult(String nextLog, Long offset) {
 }
